organize modal directory further
