+++ packages/blockchain-wallet-v4-frontend/src/modals/Generic/Confirm/helpers.js
