+++ packages/blockchain-wallet-v4-frontend/src/modals/Generic/PromptInput/index.js
