+++ packages/blockchain-wallet-v4-frontend/src/modals/Generic/index.js
@@ -1,0 +1,4 @@
+import Confirm from './Confirm'
+import PromptInput from './PromptInput'
+
+export { Confirm, PromptInput }
