+++ packages/blockchain-wallet-v4-frontend/src/modals/Settings/AutoDisconnection/template.js
