+++ packages/blockchain-wallet-v4-frontend/src/modals/Settings/index.js
@@ -1,3 +1,4 @@
+import AutoDisconnection from './AutoDisconnection'
 import ConfirmDisable2FA from './ConfirmDisable2FA'
 import SecondPassword from './SecondPassword'
 import TwoStepGoogleAuthenticator from './TwoStepGoogleAuthenticator'
@@ -5,6 +6,7 @@ import TwoStepSetup from './TwoStepSetup'
 import TwoStepYubico from './TwoStepYubico'
 
 export {
+  AutoDisconnection,
   ConfirmDisable2FA,
   SecondPassword,
   TwoStepGoogleAuthenticator,
