+++ packages/blockchain-wallet-v4-frontend/src/modals/Transactions/EditDescription/template.js
