+++ packages/blockchain-wallet-v4-frontend/src/modals/Transactions/index.js
@@ -1,0 +1,4 @@
+import EditTxDescription from './EditDescription'
+import TransactionReport from './TransactionReport'
+
+export { EditTxDescription, TransactionReport }
